Create fishingtest4.py
This was originally intended to be a definition, but it wasn't working properly . . .  It's just printing the choice and not giving the further options.  The "while" near the top was to hopefully make a loop.

My goal if this worked was to then have it go into a new definition that functioned similarly, wherein the fish did such actions as run, get tired, jump, et cetera, and you would have to react appropriately.